casts

+++ src/autlib1.c
@@ -4456,14 +4456,14 @@ cnrlbv(iap_type *iap, rap_type *rap, double *par, int64 *icp,
     nuzr = iap->nuzr;
     itpst = iap->itpst;
 
-    ups = malloc(sizeof(*ups)*(ntst + 1)*(ndim*ncol));
-    upoldp = malloc(sizeof(*upoldp)*(ntst + 1)*(ndim*ncol));
-    uoldps = malloc(sizeof(*uoldps)*(ntst + 1)*(ndim*ncol));
-    udotps = malloc(sizeof(*udotps)*(ntst + 1)*(ndim*ncol));
-    dups = malloc(sizeof(*dups)*(ntst + 1)*(ndim*ncol));
-    fa = malloc(sizeof(*fa)*(ntst + 1)*(ndim*ncol));
-    dtm = malloc(sizeof(*dtm)*(ntst + 1));
-    tm = malloc(sizeof(*tm)*(ntst + 1));
+    ups = malloc(sizeof(*ups)*(usize)((ntst + 1)*(ndim*ncol)));
+    upoldp = malloc(sizeof(*upoldp)*(usize)(ntst + 1)*(ndim*ncol));
+    uoldps = malloc(sizeof(*uoldps)*(usize)(ntst + 1)*(ndim*ncol));
+    udotps = malloc(sizeof(*udotps)*(usize)(ntst + 1)*(ndim*ncol));
+    dups = malloc(sizeof(*dups)*(usize)(ntst + 1)*(ndim*ncol));
+    fa = malloc(sizeof(*fa)*(usize)(ntst + 1)*(ndim*ncol));
+    dtm = malloc(sizeof(*dtm)*(usize)(ntst + 1));
+    tm = malloc(sizeof(*tm)*(usize)(ntst + 1));
     fc = malloc(sizeof(*fc)*(iap->nbc + iap->nint + 1));
     p0 = malloc(sizeof(*(p0))*ndim*ndim);
     p1 = malloc(sizeof(*(p1))*ndim*ndim);
